requirements.txt
pytorch==1.6.0+cu101
itertools